Move API controllers to their own namespace.
