+++ src/Controller/API/GraphController.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Controller;
+namespace App\Controller\API;
 
 use App\Entity\Device;
 use App\Entity\Probe;
